// src/app/_dto/TbldFixAss.dto.ts
export class TbldFixAss {

    DFixAss_Grid_SysID: number;
    DFixAss_AccuDep_SysID: number;
    DFixAss_AccuDep_Code: string;
    DFixAss_AccuDep_Name: string;
    DFixAss_DepSysID: number;
    DFixAss_DepCode: string;
    DFixAss_DepName: string;
    DFixAss_RemSysID: number;
    DFixAss_RemCode: string;
    DFixAss_RemName: string;

}